adding function to retrieve committers
